refactor: move devices to modules and fix recursive watchers
